+++ apps/desktop/src/main.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import { MantineProvider } from '@mantine/core';
-// import { DatesProvider } from '@mantine/dates';
 import { Notifications } from '@mantine/notifications';
 import { createRoot } from 'react-dom/client';
 
@@ -9,7 +8,6 @@ import { App } from './App';
 
 // Import Mantine CSS
 import '@mantine/core/styles.css';
-// import '@mantine/dates/styles.css';
 import '@mantine/notifications/styles.css';
 
 // Import our custom styles